[F-679] Extract duplicated utilities across Anthropic and OpenAI providers
F-583 (Anthropic) and F-584 (OpenAI) shipped near-identical helpers in
parallel. Three of those helpers were byte-identical between the two
providers:

- map_sse_error: SseError -> StreamErrorKind classifier
- truncate: UTF-8-safe prefix truncation for HTTP error logging
- ClientConfig + Default + build_stream_client: streaming reqwest::Client
  builder with the shared connect/read/keepalive timeout posture

Consolidated into a new crate-private module crates/forge-providers/src/http_util.rs
exposing HttpClientConfig, build_stream_client, map_sse_error, and truncate.
Both Anthropic and OpenAI (including the build_stream_client_default thunk
re-used by CustomOpenAiProvider) now route through it.

Ollama keeps its own ClientConfig and helpers — its public-API surface
(consumed by tests/ollama.rs) and second build_request_client both diverge
from the SSE-streaming providers and were intentionally left untouched.
